backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
langchain>=0.1.0
langgraph>=0.0.1
langchain-openai>=0.1.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
faiss-cpu>=1.8.0
numpy>=1.26.0
httpx>=0.27.0
python-dotenv>=1.0.1
stripe>=8.0.0

